Parameterize functional tests (#369)
* Introduced abstract functional test class EndToEnd

* Functional test now takes arguments

* lazy init sparks

* Edited functional test summary at the end.

* Refactored test suite execution

* Moved EndToEnd tests to a separate package

* Refactored functional test controls

* Edited comments

* Edited readme

* Close spark session before leaving

* Close spark session before leaving

* Edited readme

* Fix typo in functional-tests README

* LargeDataTests now extends from EndToEnd

* Fixed heartbeat exception when -h option is called

* Fixed failing functional tests
